feat: Add server connection
